pyproject.toml
[project]
name = "walking-on-sunshine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "fastapi>=0.116.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "spotipy>=2.25.1",
    "structlog>=25.4.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "pytest>=8.4.1",
    "ruff>=0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
main = "walking_on_sunshine.main:main"

[tool.ruff]
line-length = 120

[[tool.mypy.overrides]]
module = ["spotipy.*,structlog.*"]
ignore_missing_imports = true
